fixed my table view controller file forgot to add injection...completed MVP
