+++ gigs/gigs/View Controllers/GigsTableViewController.swift
@@ -51,9 +51,13 @@ class GigsTableViewController: UITableViewController {
 
     // In a storyboard-based application, you will often want to do a little preparation before navigation
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
+        if segue.identifier == "LoginViewModalSegue" {
+        // inject dependencies
+        if let loginVC = segue.destination as? LoginViewController {
+            loginVC.gigController = gigController
+            }
+        }
     }
-    
+
 
 }
